maro/requirements.build.txt
Cython==0.29.14
numpy==1.19.1
PyYAML==5.3.1